#Update:  - Authorization of books, authors and user apis

+++ authors/views.py
@@ -4,6 +4,7 @@ from rest_framework.response import Response
 
 from authors.forms import CreateAuthorForm, UpdateAuthorForm, DeleteAuthorForm
 from authors.service import AuthorsService
+from library_management import config
 
 
 # Create your views here.
@@ -17,12 +18,16 @@ def create_author(request):
 
         if form.is_valid():
             data = form.cleaned_data
-            # import pdb;pdb.set_trace()
-            response = AuthorsService().create_author(data)
-            if response.get("id"):
-                return Response(response, status=status.HTTP_200_OK)
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = AuthorsService().create_author(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_400_BAD_REQUEST)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
         else:
             return Response({"message": "Invalid Value"}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -41,12 +46,16 @@ def update_author(request):
 
         if form.is_valid():
             data = form.cleaned_data
-
-            response = AuthorsService().update_author(data)
-            if response.get("id"):
-                return Response(response, status=status.HTTP_200_OK)
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = AuthorsService().update_author(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_400_BAD_REQUEST)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
 
     except Exception as ex:
         return Response({"message": "Something happened wrong!", "data": ex},
@@ -61,11 +70,16 @@ def delete_author(request):
         form = DeleteAuthorForm(request.POST)
         if form.is_valid():
             data = form.cleaned_data
-            response = AuthorsService().delete_author(data)
-            if response.get("id"):
-                return Response(response, status=status.HTTP_200_OK)
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = AuthorsService().delete_author(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_400_BAD_REQUEST)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
 
     except Exception as ex:
         return Response({"message": "Something happened wrong!", "data": ex},
