+++ books/views.py
@@ -1,9 +1,10 @@
-from rest_framework import status
+from rest_framework import status, permissions
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.response import Response
 
-from books.forms import CreateBookForm, UpdateBookForm, DeleteBookForm
-from books.service import BooksService
+from books.forms import CreateBookForm, UpdateBookForm, DeleteBookForm, CreateBookLoanForm
+from books.service import BooksService, BookLoanService
+from library_management import config
 
 
 # Create your views here.
@@ -15,11 +16,16 @@ def create_book(request):
         form = CreateBookForm(request.POST)
         if form.is_valid():
             data = form.cleaned_data
-            response = BooksService().create_book(data)
-            if response.get("id"):
-                return Response(response, status=status.HTTP_200_OK)
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = BooksService().create_book(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_409_CONFLICT)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
         else:
             return Response({"message": "Invalid Value"}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -38,12 +44,16 @@ def update_book(request):
 
         if form.is_valid():
             data = form.cleaned_data
-
-            response = BooksService().update_book(data)
-            if response.get("id"):
-                return Response(response, status=status.HTTP_200_OK)
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = BooksService().update_book(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_409_CONFLICT)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
         else:
             return Response({"message": "Invalid Value"}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -60,12 +70,39 @@ def delete_book(request):
         form = DeleteBookForm(request.POST)
         if form.is_valid():
             data = form.cleaned_data
+            library_admin = request.user.groups.filter(name=config.LIBRARY_ADMIN)
+            if request.user.is_superuser or len(library_admin) > 0:
+                response = BooksService().delete_book(data)
+                if response.get("id"):
+                    return Response(response, status=status.HTTP_200_OK)
+                else:
+                    return Response(response, status=status.HTTP_409_CONFLICT)
+            else:
+                return Response({"detail": "Unauthorized Access"},
+                                status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            return Response({"message": "Invalid Value"}, status=status.HTTP_400_BAD_REQUEST)
 
-            response = BooksService().delete_book(data)
+    except Exception as ex:
+        return Response({"message": "Something happened wrong!", "data": ex},
+                        status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+
+# Create your views here.
+@api_view(['POST'])
+def create_book_loan(request):
+    try:
+        form = CreateBookLoanForm(request.POST)
+        if form.is_valid():
+
+            data = form.cleaned_data
+            response = BookLoanService().create_book_loan(data)
             if response.get("id"):
                 return Response(response, status=status.HTTP_200_OK)
             else:
-                return Response(response, status=status.HTTP_400_BAD_REQUEST)
+                return Response(response, status=status.HTTP_409_CONFLICT)
+
+
         else:
             return Response({"message": "Invalid Value"}, status=status.HTTP_400_BAD_REQUEST)
